requirements.txt
streamlit
sentence-transformers
transformers
faiss-cpu
huggingface_hub
pymupdf
torch
tqdm